Rename to Polyrific.Catapult.TaskProviders.Core
