+++ src/Engine/Polyrific.Catapult.Engine.Core/PluginManager.cs
@@ -9,7 +9,7 @@ using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using Polyrific.Catapult.TaskProvider.Core;
+using Polyrific.Catapult.TaskProviders.Core;
 using Polyrific.Catapult.Shared.Common;
 
 namespace Polyrific.Catapult.Engine.Core
