+++ src/Plugins/BuildProvider/Polyrific.Catapult.Plugins.DotNetCore/src/Polyrific.Catapult.Plugins.DotNetCore.csproj
@@ -12,7 +12,7 @@
     </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Polyrific.Catapult.TaskProvider.Core\Polyrific.Catapult.TaskProvider.Core.csproj" />
+    <ProjectReference Include="..\..\..\Polyrific.Catapult.TaskProviders.Core\Polyrific.Catapult.TaskProviders.Core.csproj" />
   </ItemGroup>
 
 </Project>
