+++ src/Plugins/BuildProvider/Polyrific.Catapult.Plugins.DotNetCore/tests/BuildProviderTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Polyrific.Catapult.TaskProvider.Core.Configs;
+using Polyrific.Catapult.TaskProviders.Core.Configs;
 using Polyrific.Catapult.Plugins.DotNetCore.Helpers;
 using Xunit;
