+++ src/Plugins/GeneratorProvider/Polyrific.Catapult.Plugins.AspNetCoreMvc/tests/CodeGeneratorProviderTests.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
-using Polyrific.Catapult.TaskProvider.Core.Configs;
+using Polyrific.Catapult.TaskProviders.Core.Configs;
 using Polyrific.Catapult.Shared.Dto.Constants;
 using Polyrific.Catapult.Shared.Dto.ProjectDataModel;
 using Xunit;
