+++ src/Plugins/HostingProvider/Polyrific.Catapult.Plugins.AzureAppService/src/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using Polyrific.Catapult.TaskProvider.Core;
+using Polyrific.Catapult.TaskProviders.Core;
 
 namespace Polyrific.Catapult.Plugins.AzureAppService
 {
