+++ src/Plugins/Polyrific.Catapult.TaskProviders.Core/CodeGeneratorProvider.cs
@@ -3,31 +3,38 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Polyrific.Catapult.TaskProvider.Core.Configs;
+using Polyrific.Catapult.TaskProviders.Core.Configs;
 using Polyrific.Catapult.Shared.Dto.Constants;
+using Polyrific.Catapult.Shared.Dto.ProjectDataModel;
 
-namespace Polyrific.Catapult.TaskProvider.Core
+namespace Polyrific.Catapult.TaskProviders.Core
 {
-    public abstract class TestProvider : TaskProvider
+    public abstract class CodeGeneratorProvider : TaskProvider
     {
-        protected TestProvider(string[] args) 
+        protected CodeGeneratorProvider(string[] args) 
             : base(args)
         {
             ParseArguments();
         }
 
-        public override string Type => PluginType.TestProvider;
+        public override string Type => PluginType.GeneratorProvider;
 
         public sealed override void ParseArguments()
         {
             base.ParseArguments();
-
+            
             foreach (var key in ParsedArguments.Keys)
             {
                 switch (key.ToLower())
                 {
+                    case "project":
+                        ProjectName = ParsedArguments[key].ToString();
+                        break;
+                    case "models":
+                        Models = JsonConvert.DeserializeObject<List<ProjectDataModelDto>>(ParsedArguments[key].ToString());
+                        break;
                     case "config":
-                        Config = JsonConvert.DeserializeObject<TestTaskConfig>(ParsedArguments[key].ToString());
+                        Config = JsonConvert.DeserializeObject<GenerateTaskConfig>(ParsedArguments[key].ToString());
                         break;
                     case "additional":
                         AdditionalConfigs = JsonConvert.DeserializeObject<Dictionary<string, string>>(ParsedArguments[key].ToString());
@@ -43,27 +50,27 @@ namespace Polyrific.Catapult.TaskProvider.Core
             switch (ProcessToExecute)
             {
                 case "pre":
-                    var error = await BeforeTest();
+                    var error = await BeforeGenerate();
                     if (!string.IsNullOrEmpty(error))
                         result.Add("errorMessage", error);
                     break;
                 case "main":
-                    (string testResultLocation, Dictionary<string, string> outputValues, string errorMessage) = await Test();
-                    result.Add("testResultLocation", testResultLocation);
+                    (string outputLocation, Dictionary<string, string> outputValues, string errorMessage) = await Generate();
+                    result.Add("outputLocation", outputLocation);
                     result.Add("outputValues", outputValues);
                     result.Add("errorMessage", errorMessage);
                     break;
                 case "post":
-                    error = await AfterTest();
+                    error = await AfterGenerate();
                     if (!string.IsNullOrEmpty(error))
                         result.Add("errorMessage", error);
                     break;
                 default:
-                    await BeforeTest();
-                    (testResultLocation, outputValues, errorMessage) = await Test();
-                    await AfterTest();
+                    await BeforeGenerate();
+                    (outputLocation, outputValues, errorMessage) = await Generate();
+                    await AfterGenerate();
 
-                    result.Add("testResultLocation", testResultLocation);
+                    result.Add("outputLocation", outputLocation);
                     result.Add("outputValues", outputValues);
                     result.Add("errorMessage", errorMessage);
                     break;
@@ -73,9 +80,19 @@ namespace Polyrific.Catapult.TaskProvider.Core
         }
 
         /// <summary>
-        /// Test task configuration
+        /// Name of the project
         /// </summary>
-        public TestTaskConfig Config { get; set; }
+        public string ProjectName { get; set; }
+
+        /// <summary>
+        /// Project data models
+        /// </summary>
+        public List<ProjectDataModelDto> Models { get; set; }
+
+        /// <summary>
+        /// Generate task configuration
+        /// </summary>
+        public GenerateTaskConfig Config { get; set; }
 
         /// <summary>
         /// Additional configurations for specific provider
@@ -83,25 +100,25 @@ namespace Polyrific.Catapult.TaskProvider.Core
         public Dictionary<string, string> AdditionalConfigs { get; set; }
 
         /// <summary>
-        /// Process to run before executing test
+        /// Process to run before executing the code generation
         /// </summary>
         /// <returns></returns>
-        public virtual Task<string> BeforeTest()
+        public virtual Task<string> BeforeGenerate()
         {
             return Task.FromResult("");
         }
 
         /// <summary>
-        /// Run test scenario
+        /// Generate code from data models
         /// </summary>
         /// <returns></returns>
-        public abstract Task<(string testResultLocation, Dictionary<string, string> outputValues, string errorMessage)> Test();
+        public abstract Task<(string outputLocation, Dictionary<string, string> outputValues, string errorMessage)> Generate();
 
         /// <summary>
-        /// Process to run after executing test
+        /// Process to run after executing code generation
         /// </summary>
         /// <returns></returns>
-        public virtual Task<string> AfterTest()
+        public virtual Task<string> AfterGenerate()
         {
             return Task.FromResult("");
         }
