+++ src/Plugins/Polyrific.Catapult.TaskProviders.Core/TaskProvider.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Polyrific.Catapult.Shared.Dto.Constants;
 
-namespace Polyrific.Catapult.TaskProvider.Core
+namespace Polyrific.Catapult.TaskProviders.Core
 {
     public abstract class TaskProvider
     {
