+++ src/Plugins/TestProvider/Polyrific.Catapult.Plugins.DotNetCoreTest/src/Polyrific.Catapult.Plugins.DotNetCoreTest.csproj
@@ -12,7 +12,7 @@
     </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Polyrific.Catapult.TaskProvider.Core\Polyrific.Catapult.TaskProvider.Core.csproj" />
+    <ProjectReference Include="..\..\..\Polyrific.Catapult.TaskProviders.Core\Polyrific.Catapult.TaskProviders.Core.csproj" />
   </ItemGroup>
 
 </Project>
