+++ src/Plugins/TestProvider/Polyrific.Catapult.Plugins.DotNetCoreTest/src/Program.cs
@@ -3,7 +3,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using Polyrific.Catapult.TaskProvider.Core;
+using Polyrific.Catapult.TaskProviders.Core;
 
 namespace Polyrific.Catapult.Plugins.DotNetCoreTest
 {
